+++ src/main/java/com/jinnova/smartpad/db/DaoSupport.java
@@ -94,8 +94,8 @@ class DaoSupport {
 	static final String GPS_FIELDS = "gps_lon=?, gps_lat=?, gps_inherit=?";
 
 	static void populateGps(ResultSet rs, GPSInfo gps) throws SQLException {
-		gps.setLongitude(rs.getFloat("gps_lon"));
-		gps.setLatitude(rs.getFloat("gps_lat"));
+		gps.setLongitude(rs.getBigDecimal("gps_lon"));
+		gps.setLatitude(rs.getBigDecimal("gps_lat"));
 		
 		//must be after setting gps longitude/latitude
 		gps.setInheritFrom(rs.getString("gps_inherit"));
@@ -103,8 +103,8 @@ class DaoSupport {
 	}
 
 	static int setGpsFields(PreparedStatement ps, GPSInfo gps, int i) throws SQLException {
-		ps.setFloat(i++, gps.getLongitude());
-		ps.setFloat(i++, gps.getLatitude());
+		ps.setBigDecimal(i++, gps.getLongitude());
+		ps.setBigDecimal(i++, gps.getLatitude());
 		ps.setString(i++, gps.getInheritFrom());
 		gps.clearModifiedFlag();
 		return i;
